refactor(portal): move shared fetch-apps query to scenes/common/apps
Both the v2 AppSelector and the v3 AppsDropdown query the team's apps, so
the FetchApps doc moves to scenes/common/apps for a single shared copy.
Pure relocation (git mv, byte-identical) plus import repoints across the 4
consumers and the AppsDropdown test mock. The separate Team/page/Apps
fetch-apps doc is unrelated and untouched.
